tests/wpt/web-platform-tests/html/canvas/offscreen/transformations/2d.transformation.scale.3d.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.transformation.scale.3d</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.transformation.scale.3d</h1>
<p class="desc">scale() function with three arguments modifies the underlying matrix appropriately</p>


<script>
var t = async_test("scale() function with three arguments modifies the underlying matrix appropriately");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

const sx = 2;
const sy = 3;
const sz = 4;
ctx.scale(sx, sy, sz);
let canvasTransform = ctx.getTransform();
_assert(canvasTransform.m11 = sx, "canvasTransform.m11 = sx");
_assert(canvasTransform.m22 = sy, "canvasTransform.m22 = sy");
_assert(canvasTransform.m33 = sz, "canvasTransform.m33 = sz");
ctx.scale(sx, sy, sz);
canvasTransform = ctx.getTransform();
_assert(canvasTransform.m11 = 2 * sx, "canvasTransform.m11 = 2 * sx");
_assert(canvasTransform.m22 = 2 * sy, "canvasTransform.m22 = 2 * sy");
_assert(canvasTransform.m33 = 2 * sz, "canvasTransform.m33 = 2 * sz");
t.done();

});
</script>
